exercise-09

+++ src/main/java/basic/Exer09.java
@@ -3,10 +3,10 @@ package basic;
 public class Exer09 {
     public double calPi(double epsilon) {
         double result = 0;
-        double i = 2;
+        int i = 2;
         double temp = 0;
 
-        while ((temp = 4 / (i * (i + 1) * (i + 2))) >= epsilon) {
+        while ((temp = 4.0 / (i * (i + 1) * (i + 2))) >= epsilon) {
             if (i % 4 != 0) {
                 result += temp/4;
